# cmd/swag/docs/swagger.yaml
basePath: /v1/2025
definitions:
  model.LoginInput:
    properties:
      user_account:
        type: string
      user_password:
        type: string
    type: object
  model.RegisterInput:
    properties:
      verify_key:
        type: string
      verify_purpose:
        type: string
      verify_type:
        type: integer
    type: object
  model.UpdatePasswordRegisterInput:
    properties:
      user_password:
        type: string
      user_token:
        type: string
    type: object
  model.VerifyInput:
    properties:
      verify_code:
        type: string
      verify_key:
        type: string
    type: object
  response.ErrorResponseData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8888
info:
  contact:
    email: official.buiminhhoat@gmail.com
    name: Hoat Bui Minh
    url: https://github.com/buiminhhoat/go-ecommerce-backend-api
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/buiminhhoat/go-ecommerce-backend-api
  title: API Documentation Ecommerce Backend SHOPDEVGO
  version: 1.0.0
paths:
  /user/login:
    post:
      consumes:
      - application/json
      description: User Login
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: User Login
      tags:
      - account management
  /user/register:
    post:
      consumes:
      - application/json
      description: When user is registered send otp to end
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: User Registration
      tags:
      - account management
  /user/update_password_register:
    post:
      consumes:
      - application/json
      description: UpdatePasswordRegister
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePasswordRegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: UpdatePasswordRegister
      tags:
      - account management
  /user/verify_account:
    post:
      consumes:
      - application/json
      description: Verify OTP Login By User
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.VerifyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Verify OTP Login By User
      tags:
      - account management
schemes:
- http
swagger: "2.0"